testing/web-platform/tests/2dcontext/drawing-images-to-the-canvas/drawimage_html_image_6_srgb.html

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<canvas id="dest" height="100" width="100"></canvas>

<script>
var sourceWidth = 100;
var sourceHeight = 100;
var smoothingEnabled = false;
var destCanvas = document.getElementById('dest');
var sourceImg = document.createElement('img');
sourceImg.src = '../2x2-srgb.png'
sourceImg.width = sourceWidth;
sourceImg.height = sourceHeight;
var destCtx = destCanvas.getContext('2d');

function checkPixel(x, y, expected, epsilon) {
    var actual = destCtx.getImageData(x, y, 1, 1).data;
    assert_true(actual.length === expected.length);
    for (var i=0; i < actual.length; i++)
        assert_approx_equals(actual[i], expected[i], epsilon);
}

var pixelTests = [
    ['Pixel Top Right Canvas Background', 0, 99, [255, 0, 0, 255], 1],
    ['Pixel Bottom Left Canvas Background', 99, 0, [255, 0, 0, 255], 1],
    ['Pixel Bottom Right Canvas Background', 99, 99, [255, 0, 0, 255], 1],

    ['Pixel Surrounding Top Left Canvas Background', 0, 32, [255, 0, 0, 255], 1],
    ['Pixel Surrounding Bottom Left Canvas Background', 32, 0, [255, 0, 0, 255], 1],
    ['Pixel Surrounding Orthogonal Canvas Background', 32, 32, [255, 0, 0, 255], 1],

    ['Pixel Top Left Image', 0, 0, [253, 140, 245, 255], 1],
    ['Pixel Top Right Image', 0, 31, [253, 140, 245, 255], 1],
    ['Pixel Bottom Left Image', 31, 0, [253, 140, 245, 255], 1],
    ['Pixel Bottom Right Image', 31, 31, [253, 140, 245, 255], 1],
];

function runTest() {
    destCtx.fillStyle = "#FF0000";
    destCtx.fillRect(0, 0,  destCanvas.width, destCanvas.height);
    destCtx.imageSmoothingEnabled = smoothingEnabled;
    // 8 arguments, both destination and source origins are 0, 0
    // An area of 32x32 pixels of the source image will be copied to
    // an area of 32x32 pixels of the destination canvas
    destCtx.drawImage(sourceImg, 0, 0, 32, 32, 0, 0, 32, 32);
    generate_tests(checkPixel, pixelTests);
}

async_test(t => {
        window.onload = function() {
            t.step(runTest);
            t.done();
        }
}, 'Draw a 32x32 area of a 100x100 image to a 32x32 area of a 100x100 canvas at 0,0.');
</script>
